Move validation to the API handler
